+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/PlayerHeaderFragment.kt
@@ -17,11 +17,13 @@ import androidx.core.view.GestureDetectorCompat
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import au.com.shiftyjelly.pocketcasts.models.entity.Episode
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.models.to.Chapter
 import au.com.shiftyjelly.pocketcasts.models.type.EpisodeStatusEnum
 import au.com.shiftyjelly.pocketcasts.player.R
 import au.com.shiftyjelly.pocketcasts.player.databinding.AdapterPlayerHeaderBinding
+import au.com.shiftyjelly.pocketcasts.player.view.ShelfFragment.Companion.AnalyticsProp
 import au.com.shiftyjelly.pocketcasts.player.view.video.VideoActivity
 import au.com.shiftyjelly.pocketcasts.player.viewmodel.PlayerViewModel
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
@@ -70,6 +72,7 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
     @Inject lateinit var playbackManager: PlaybackManager
     @Inject lateinit var settings: Settings
     @Inject lateinit var warningsHelper: WarningsHelper
+    @Inject lateinit var analyticsTracker: AnalyticsTrackerWrapper
 
     lateinit var imageLoader: PodcastImageLoaderThemed
     private val viewModel: PlayerViewModel by activityViewModels()
@@ -118,6 +121,7 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
         binding.seekBar.changeListener = object : PlayerSeekBar.OnUserSeekListener {
             override fun onSeekPositionChangeStop(progress: Int, seekComplete: () -> Unit) {
                 viewModel.seekToMs(progress, seekComplete)
+                playbackManager.trackPlaybackSeek(progress, PlaybackSource.PLAYER)
             }
 
             override fun onSeekPositionChanging(progress: Int) {}
@@ -155,15 +159,25 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
         binding.share.setOnClickListener { onShareClick() }
         binding.playerActions.setOnClickListener { onMoreClicked() }
         binding.podcast.setOnClickListener { showPodcast() }
-        binding.played.setOnClickListener { viewModel.markCurrentlyPlayingAsPlayed(requireContext())?.show(childFragmentManager, "mark_as_played") }
-        binding.archive.setOnClickListener { viewModel.archiveCurrentlyPlaying(resources)?.show(childFragmentManager, "archive") }
-        binding.download.setOnClickListener { viewModel.downloadCurrentlyPlaying() }
+        binding.played.setOnClickListener {
+            trackShelfAction(ShelfItem.Played.analyticsValue)
+            viewModel.markCurrentlyPlayingAsPlayed(requireContext())?.show(childFragmentManager, "mark_as_played")
+        }
+        binding.archive.setOnClickListener {
+            trackShelfAction(ShelfItem.Archive.analyticsValue)
+            viewModel.archiveCurrentlyPlaying(resources)?.show(childFragmentManager, "archive")
+        }
+        binding.download.setOnClickListener {
+            trackShelfAction(ShelfItem.Download.analyticsValue)
+            viewModel.downloadCurrentlyPlaying()
+        }
         binding.videoView.playbackManager = playbackManager
         binding.videoView.setOnClickListener { onFullScreenVideoClick() }
 
         CastButtonFactory.setUpMediaRouteButton(binding.root.context, binding.castButton)
         binding.castButton.setAlwaysVisible(true)
         binding.castButton.updateColor(ThemeColor.playerContrast03(theme.activeTheme))
+        binding.castButton.setOnClickListener { trackShelfAction(ShelfItem.Cast.analyticsValue) }
 
         setupUpNextDrag(view, binding.topView)
 
@@ -171,7 +185,7 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
             val headerViewModel = it.podcastHeader
             binding.viewModel = headerViewModel
 
-            binding.largePlayButton.setPlaying(isPlaying = headerViewModel.isPlaying, animate = false)
+            binding.largePlayButton.setPlaying(isPlaying = headerViewModel.isPlaying, animate = true)
 
             val playerContrast1 = ThemeColor.playerContrast01(headerViewModel.theme)
             binding.largePlayButton.setCircleTintColor(playerContrast1)
@@ -426,22 +440,27 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
     }
 
     override fun onEffectsClick() {
+        trackShelfAction(ShelfItem.Effects.analyticsValue)
         EffectsFragment().show(parentFragmentManager, "effects_sheet")
     }
 
     override fun onSleepClick() {
+        trackShelfAction(ShelfItem.Sleep.analyticsValue)
         SleepFragment().show(parentFragmentManager, "effects_sheet")
     }
 
     override fun onStarClick() {
+        trackShelfAction(ShelfItem.Star.analyticsValue)
         viewModel.starToggle()
     }
 
     override fun onShareClick() {
+        trackShelfAction(ShelfItem.Share.analyticsValue)
         viewModel.shareDialog(context, childFragmentManager)?.show()
     }
 
     private fun showPodcast() {
+        trackShelfAction(ShelfItem.Podcast.analyticsValue)
         val podcast = viewModel.podcast
         (activity as FragmentHostListener).closePlayer()
         if (podcast != null) {
@@ -452,17 +471,13 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
     }
 
     override fun onPreviousChapter() {
+        analyticsTracker.track(AnalyticsEvent.PLAYER_PREVIOUS_CHAPTER_TAPPED)
         viewModel.previousChapter()
     }
 
     override fun onNextChapter() {
+        analyticsTracker.track(AnalyticsEvent.PLAYER_NEXT_CHAPTER_TAPPED)
         viewModel.nextChapter()
-    }
-
-    override fun onPlayingEpisodeActionsClick() {
-        viewModel.episode?.let { episode ->
-            (activity as? FragmentHostListener)?.openEpisodeDialog(episode.uuid, (episode as? Episode)?.podcastUuid, forceDark = true)
-        }
     }
 
     fun onMoreClicked() {
@@ -470,6 +485,7 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
         if (childFragmentManager.fragments.firstOrNull() is ShelfBottomSheet) {
             return
         }
+        analyticsTracker.track(AnalyticsEvent.PLAYER_SHELF_OVERFLOW_MENU_SHOWN)
         ShelfBottomSheet().show(childFragmentManager, "shelf_bottom_sheet")
     }
 
@@ -524,5 +540,12 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
 
     override fun onHeaderChapterClick(chapter: Chapter) {
         (parentFragment as? PlayerContainerFragment)?.openChaptersAt(chapter)
+    }
+
+    private fun trackShelfAction(analyticsAction: String) {
+        analyticsTracker.track(
+            AnalyticsEvent.PLAYER_SHELF_ACTION_TAPPED,
+            mapOf(AnalyticsProp.Key.FROM to AnalyticsProp.Value.SHELF, AnalyticsProp.Key.ACTION to analyticsAction)
+        )
     }
 }