+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/ProfileEpisodeListFragment.kt
@@ -18,6 +18,7 @@ import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.models.entity.Episode
 import au.com.shiftyjelly.pocketcasts.models.entity.Playable
+import au.com.shiftyjelly.pocketcasts.models.type.EpisodeViewSource
 import au.com.shiftyjelly.pocketcasts.podcasts.R
 import au.com.shiftyjelly.pocketcasts.podcasts.databinding.FragmentProfileEpisodeListBinding
 import au.com.shiftyjelly.pocketcasts.podcasts.view.components.PlayButton
@@ -99,7 +100,12 @@ class ProfileEpisodeListFragment : BaseFragment(), Toolbar.OnMenuItemClickListen
 
     val onRowClick = { episode: Playable ->
         if (episode is Episode) {
-            val fragment = EpisodeFragment.newInstance(episode)
+            val episodeViewSource = when (mode) {
+                Mode.Downloaded -> EpisodeViewSource.DOWNLOADS
+                Mode.History -> EpisodeViewSource.LISTENING_HISTORY
+                Mode.Starred -> EpisodeViewSource.STARRED
+            }
+            val fragment = EpisodeFragment.newInstance(episode = episode, source = episodeViewSource)
             fragment.show(parentFragmentManager, "episode_card")
         }
     }
@@ -183,10 +189,12 @@ class ProfileEpisodeListFragment : BaseFragment(), Toolbar.OnMenuItemClickListen
             binding?.toolbar?.isVisible = !isMultiSelecting
             binding?.multiSelectToolbar?.setNavigationIcon(R.drawable.ic_arrow_back)
 
-            if (isMultiSelecting) {
-                trackMultiSelectEntered()
-            } else if (wasMultiSelecting) {
-                trackMultiSelectExited()
+            if ((activity as? FragmentHostListener)?.isUpNextShowing() == false) {
+                if (isMultiSelecting) {
+                    trackMultiSelectEntered()
+                } else if (wasMultiSelecting) {
+                    trackMultiSelectExited()
+                }
             }
 
             adapter.notifyDataSetChanged()