+++ modules/services/ui/src/main/java/au/com/shiftyjelly/pocketcasts/ui/helper/FragmentHostListener.kt
@@ -2,6 +2,7 @@ package au.com.shiftyjelly.pocketcasts.ui.helper
 
 import android.view.View
 import androidx.fragment.app.Fragment
+import au.com.shiftyjelly.pocketcasts.models.type.EpisodeViewSource
 
 interface FragmentHostListener {
     fun addFragment(fragment: Fragment, onTop: Boolean = false)
@@ -22,8 +23,9 @@ interface FragmentHostListener {
     fun updateStatusBar()
     fun updatePlayerView()
     fun getPlayerBottomSheetState(): Int
-    fun openEpisodeDialog(episodeUuid: String?, podcastUuid: String?, forceDark: Boolean)
+    fun openEpisodeDialog(episodeUuid: String?, source: EpisodeViewSource, podcastUuid: String?, forceDark: Boolean)
     fun lockPlayerBottomSheet(locked: Boolean)
     fun updateSystemColors()
     fun overrideNextRefreshTimer()
+    fun isUpNextShowing(): Boolean
 }